ai-agent/pyproject.toml
[project]
name = "kalshi-ai-agent"
version = "0.1.0"
description = "AI agent that uses MCP to interact with Kalshi markets"
requires-python = ">=3.10"

dependencies = [
    "mcp>=1.6.0",
    "ollama>=0.1.0",
    "python-dotenv>=1.0.0",
    "asyncio",
]

[project.optional-dependencies]
dev = [
    "black",
    "ruff",
    "mypy"
]

[project.scripts]
agent = "agent:main"

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "C90", "I", "N", "UP", "ANN", "ASYNC", "S", "B"]
ignore = ["ANN101", "ANN102", "S101"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
